fix: deployment preview  error
